Update RoundState.proceed() & implement BoardState.proceed()
RoundState.proceed() now handles NUM_BOARDS actions at a time, applying them to non-settled boards, so that each board can proceed with a single action as RoundState used to.

Makes better use of OOP, and adds "settled" attribute to BoardState to identify when individual Boards are ready for the next street.